Graphics: - Added animations - Exposed animations to Lua - Renderer is now multithreading friendly - Cleaned up debug renderer - Fixed frustum culling
Networking:
- Player creation and destroying

Physics:
- Fixed issue where physics did not always sync before simulation
- Constraint fixes:
-- Creation / deletion
-- Ability to attach / detach entities
-- Updated interface / added functions
-- Fixed hinge constraint implementation
- Fixed kinematic body issue
- Changed class hierarchy so a public interface is defined for each shape
- Fixed capsule and cylinder colliders jittering when their radius is modified
- Minor bugfixes

Rewinder:
- Added light system rewinding
- Added camera system rewinding

Editor:
- Added inspector for multiple properties
- Added shader data class
- Fixed binary reader not loading certain types properly
- Added color picker
- Finalized sulphur projects (still needs work on engine side)

Builder:
- Made pipelines more project friendly
- Improved error messages
- Fixed small bugs

Generator:
- You can now refresh the banners

Misc:
- Multi-processor compilation support
- Bug fixes and stability improvements

+++ src/foundation/pipeline-assets/shader.h
@@ -110,8 +110,8 @@ namespace sulphur
       ConcreteType concrete_type; //!< concrete type of the resource. the actual data type of the resource
       uint32_t cols;  //!< resource width ie number of rows in a matrix
       uint32_t vec_size; //!< resource height ie. number of cols in a matrix
-      size_t offset; //!< memory offset in bytes
-      size_t size; //!< size of the resource in bytes
+      uint64_t offset; //!< memory offset in bytes
+      uint64_t size; //!< size of the resource in bytes
       Vector<ShaderResource> members; //!< member fields of a resource i.e. members of a struct
     };
